Tank Movement
A single Tank is now working, though the script still needs some work.
Occasionally, the tanks have an issue with rotating, but it's about 1/20
of the time.

Next, will work on spawning multiple tanks at given positions. Might use
a text file to determine each of the tank spawn locations, or might look
for the tank waypoints.