Dev 10
Completed work:
1. Finished Enemy navigation. Turns out, Unity has a built-in navigation
system. While it means I won't be using the MultiplayerTankController
script the way I had intended, this makes the dev go a lot faster.
Still, the tanks don't move as I'd hope, but at least it's working.

2. Tanks patrol. Now, when tanks are in patrol mode, they ask the
MultiplayerSceneController for a new destination once they've reached
their previous destination. The MultiplayerSceneController passes out a
new destination based on the current and projected spread of enemy
tanks, seeking to ensure that there are 4 tanks per level. This looks
like it may have bugs as well, and needs testing.

3. Win-Lose system set up. Now, when a player takes out the other
player, a signal is sent to the MultiplayerSceneController to declare
the winner and stop the game. This is supposed to stop all movement, but
the enemy tanks continue to move for some reason.

4. Game reloads. After a game has been won, the players are prompted to
begin a new game. This reloads the current scene. There were initially
some issues with the lighting. After some research, it was determined
that this was due to continuous lighting. While a quickfix was
implemented, it still needs to be tested to ensure it's doing what it's
supposed to do.

What needs to be done:
1. Test, test, test
2. Debug, debug, debug
3. Maybe add in some particles and audio for a little bit of flair
NOTE: A last-minute build found that the end-game mechanism was keying
off of the firing-tank's collider and not the tank it was hitting.